fix: serac spack pkg: ' install' -> 'install'
This commit fixes a bug in the `phases` list in the Serac Spack
package. In this list, the last element was ' install' and has been
changed in this commit to 'install'. The leading space causes an error
when Serac clients attempt to install Serac using this Spack package,
and deleting this leading space resolves this error.