quickstart.rst: suggest python3 instead of python
This commit changes the `python` commands in the quickstart to
`python3`. On some systems, including `quartz.llnl.gov`, `python` maps
to Python 2 in default configurations instead of Python 3. I'm not
sure this change affects those commands at the moment, but it would be
more consistent with the recommended WSL/Ubuntu configurations, which
suggest installing `python3` versions of Debian packages.