serac spack pkg: add +asan variant
This commit adds a `+asan` variant to the Serac Spack package to
expose more readily the `ENABLE_ASAN` option to users building Serac
through Spack and Uberenv. This change is probably more convenient for
downstream users of Serac that build it as a dependency of their
projects via Spack/Uberenv.

Resolves #531.