Remove dimensionally dependent material model implementations
Several material models were template parameterized on the spatial
dimensionality (2 or 3), with the implementation changing based on the
dimensionality. The 2D versions were incorrect, being neither plane
strain nor plane stress.

Now: spatial dimension templating of class removed. Parameterize
operator() on spatial dimension, such that in 2D plane strain is
effected (for the in-plane components of the stress).