Switch time integration scheme to Newmark, which seems to work
The default AverageAcceleration integrator is not getting the
right answer. The name is confusing - it is a variation of the HHT
integrator, not the classic Newmark method that goes by the
"average acceleration" name.

I don't know yet if the problem with AverageAcceleration is a bug
in its implementation, or if we are calling it incorrectly.