Go back to applying traction on all surfaces as it works for both 2D and 3D
Until we have a way to paint the boundaries, there is no good way to
restrict the traction application to only the complement of the
essential boundary condition surfaces. I've left in the old code
that does this filtering by the normal direction, so that it can
be resurrected easily to get reaction forces for debugging. I
could have a really ugly conditional for 2D and 3D. I'd rather not,
since it is brittle in that it depends on the particular mesh file
and the choice of essential/natural BC surfaces.