Add residual convergence criterion to scalar solver
This lets us set the tolerance for the yield condition (obviously
stress-based), which determines whether we do a radial return,
the same as the tolerance for exiting the radial return. In other
words, this lets us ensure that the updated state variables would
count as converged if we were to call the constitutive algorithm
again with those values.