Rename small strain plasticity model now that there is only one
J2Nonlinear -> J2SmallStrain
The "nonlinear" descriptor was unneeded, since now there is only
one small strain J2 model, and whether it is linear or not is up
to the user when they choose the hardening.

The old name also failed to point out that the formulation was
only for small strain theory. The rest of the code acts like
finite deformation is the default choice, (which I agree with),
so this model should point out its difference from that norm.