Change tactics, try to use custo domain integral in thermal solver for internal heat source
Using a material doesn't work, because the source term contribution
is meant to be the heat capacity, and hence gets multiplied by the
temperature rate in the residual. With a custom integral, we can
hopefully plug the internal heat source into the appropriate slot
in functional and get the right result.